Add more information to Privacy Policy to comply with GDPR.
This adds additional information to our privacy policy in order to
comply with the EU GDPR directive and adds useful descriptions of what
data we store, for how long, what reasons and more.

This also clarifies that we try our best to limit the amount
of personal information we have and to protect our users’ privacy
by design.