Initial repository setup and content transfer.
This includes files necessary for the repository workflow and
important basic files carrying information such as the licensing
and documentation for this repository.

This includes the initial Terms of Service and Privacy Policy with the
status of when Colloq went live on November, 14th 2017.

From here, all changes to the Terms of Service and Privacy Policy
are trackable via the git history.